Add `collectContextOnly` option to `RenameStrategy`
This commit introduces the `collectContextOnly` field to the `RenameStrategy` allowing users to only fetch the context information without generating name suggestions. Additionally, the 'getSuggestions' method of the `SuggestionsProvider` now expects this option as a parameter. The error handling in the `LLMRenameProvider` has also been improved to generate a default `NameSuggestionContext` in case of failure.

GitOrigin-RevId: e9d69b07f3b0b208e286f54ae7235365b9182eda