IJPL-156236 fix update info propagation in UpdateChecker
If a scheduled update check has found a platform update, the information about the plugin updates was calculated for the new version and passed into the UpdateSettingsEntryPointActionProvider without the information about the platform update. Because of this, it was possible to do the plugin update without the corresponding platform update and thus make the plugins incompatible with the current version.

GitOrigin-RevId: 0589eb2558ba9425f664a368b17e1197608e0ca4