IJPL-148224 Add theme colors for license dialog
Since the new mockups use specific colors in places
such as the terms and conditions and license details,
we need a place to put those.

The colors for the Dark theme were taken from the mockups,
the Light theme colors were derived from them in a symmetric
fashion. We have 14 shades of gray, so if we split them into
the dark half 1-7 and the light half 8-14, then 7 corresponds to 8,
6 corresponds to 9 and so on.

No colors are defined for the Classic UI for now, but the existing
ones don't look that horrible either.

GitOrigin-RevId: 7486079b26b6f0fc497d3f7cdae89a27e34e9a6f