check if caller was canceled first, then throw `CancellationException` when message bus is disposed
In the concurrent cancellation model, the coroutines are able to observe the disposed message bus,
and they shall finish ASAP. This is similar to `await` on some deferred: once it's canceled,
all clients are resumed with `CancellationException` even if they were not canceled.

This change makes "Already Disposed" exceptions treatable as regular cancellations and not failures,
because they are technically control flow exceptions.

GitOrigin-RevId: ad44cd0e9891fa1968011a2559df917ba67cfd4b