IJPL-149829 Fix classloading of Microba UI implementation classes
Since this library is not a part of the IntelliJ platform core, its classes are loaded using separate classloader.
But when UIDefaults tries to get the UI class by its FQN, it is using platform classloader and fails to find it.
The solution is to specify specific classes right when patching the UIDefaults.
But there is a downside: UI classes will be loaded eagerly instead of when a specific class is requested.

GitOrigin-RevId: e596f2d594f10bbcf2aa54096eed6006868614f1