[json] IJPL-63554 Introduced the if-then-else resolve-tree building  operation
To be able to work with if-else expressions met in the middle of the resolve path and to avoid various hacks related to inability to select the correct branch resulting frequently in incorrect validation results, the new IfThenElseBranchOperation is introduced.

It is responsible for choosing the right branch to resolve to and actually works during resolve phase, not during the resolved object consumption by any client extension like completion contributor or inspection.

To ensure correct branch selection, one must provide an inspected PSI element wrapper. Schema will try to process its content and select the branch depending on it. There is also a mode where all branches are considered valid - useful for completion purposes

GitOrigin-RevId: 103c7c0e97bebe1469e85b7b914be5264b4ca3af