IJPL-797 `intellij.platform.ide.util.io` review internal API: `ExecUtil` has public and internal methods
Everything related to running `sudo` is marked as internal.

Also, `readFirstLine` is marked as internal, because it's a dangerous function. Calling `readFirstLine` may potentially lead to hung processes if such a process writes a lot of text in stdout/stderr.

`escapeUnixShellArgument` is also marked as deprecated because it does its job incorrectly.

GitOrigin-RevId: a9b70cd482375be55b2e1e9e0e2f3e587a310430