[kotlin] J2K: migrate J2KNullityInferrer to use `PsiType`s directly
This change is needed for the upcoming generics nullability support.

1. Operating on `PsiType`s directly is not always correct
because a PsiType object can sometimes be recreated (for example, by PsiNewExpressionImpl.getType()),
and in that case we won't find it in our IdentityHashMap.
We should try to migrate this logic to PsiTypeElements, which are more persistent.

2. The changes in testdata related to varargs are expected:
In Kotlin, `vararg objs: Any?` means that `objs` is a not-null Array of `Any?` elements.
So, the nullability of the underlying JKJavaArrayType should be discarded,
like it was always done in JavaToJKTreeBuilder before the introduction
of J2KNullityInferrer. Since we don't infer nullability of vararg components,
it becomes nullable. The previous testdata was incorrect.

3. The change also revealed a problem with C-style array declarations,
which is mentioned in `PsiTypeElement#getType` JavaDoc. This corner case needs
to be supported for K2 nullability separately (a possible solution is to run
a Java pre-processing inspection to convert such declarations to Java style).

KTIJ-29147

GitOrigin-RevId: e32e9e069209f27675a68cb9613c23d710826cc1