PY-60597 Set global debugger to `None` and clear thread local info when stopping tracing
The global debugger reference should be set to `None` right away. Otherwise, the reference to the stopped debugger can leak to `ThreadTracer` leading to skipped breakpoints. The same applies for the thread local info. If it is not reset, the frame evaluation code will continue using a tracing function with the stale debugger reference.

GitOrigin-RevId: 62b38316e4da760213810a55aa98db289ce45720