IJPL-797 `intellij.platform.jps.model` review internal API
Mark as internal members which aren't supposed to be used by plugins. In particular, methods which modify the project model are excluded from API because they are supposed to be called from the implementation code when loading the project configuration from disk. Also, their presense would complicate migration of the JPS model's implementation to the workspace model (IJPL-409).

GitOrigin-RevId: fa717a689aba627a0a4e5bd768413447b8fc574b