[RDCT] Fix JSON Settings test can't find EditorSettings
The problem is that Json Settings test uses `componentManager.processAllImplementationClasses` API to iterate through all services registered in this component manager. But it uses instance holders inside, and they are not created for per-client services.
This commit adds `EditorSettingsExternalizable` and `CodeInsightSettings` into a hardcoded list for now, but it should be fixed properly in the future.

IJ-MR-130819

GitOrigin-RevId: 12aa05de0ddc24091e2cabf1c0ddf07e43237109