[Workspace Model] [IJPL-149482] Fork-join loading of ModuleBridge
This change was reverted in d6969bfef67fa0e629b049424876ff421d5522d2 because of IJPL-149805

To avoid the issue IJPL-149805 in a proper way, we initially finish all loadings of the bridges and only then add bridges to the external maps. This approach is similar to  `ModuleManagerBridgeImpl.loadModules`

Also, the code was updated to initialize components in a non-blocking way, also the same it happens in ` ModuleManagerBridgeImpl.loadModules`.

GitOrigin-RevId: b11709e76db12dcabb426c131f258035f5216dd4