[rdct] use proper paths to settings and caches directories when connecting to Community editions (RDCT-1430)
We use the same distribution of the frontend part for community and ultimate/professional editions of IntelliJ IDEA and PyCharm, and 'idea.path.selector' property is set to the value from the ultimate/professional edition. To use proper value when connecting to a backend from the community edition, we need to adjust it based on the product code specified in the join link in the command line arguments.

GitOrigin-RevId: 663f40075eef089d7a8e36ec0ee2a8ba5f39af10