IJPL-148224 Introduce separate keys for license dialog separator colors
We have two separators now: one in the Manage Licenses dialog,
and one in the license list. They have different colors, so we need
two keys. The previously existing LicenseDialog.separatorColor key
now corresponds to the separators in the license dialog itself.
And the new LicenseDialog.LicenseList.separatorColor key
is what LicenseDialog.separatorColor used to mean: the separators
in the license list.

GitOrigin-RevId: f7a032e8a9526676788af243ba60db9bb286003a