Refactor KotlinCodeGenerationVisitor for body expressions
The method for generating code in the KotlinCodeGenerationVisitor has been updated. The way function body expressions are handled has been modified, allowing for a clearer, more specific syntax and better error handling. Notably, the body expression is now expected to be a block, and the check for a non-empty body has been simplified.

GitOrigin-RevId: 373d91f4e47d3b93970927d741f7a4fc8dc2004f