IJPL-149878 IJent WSL FS: workarounds for cases when VM option can't be removed
In production, there are two VM Options files: the bundled one and the user-defined one.

`VMOptions.setOption` always changes the user-defined settings. It's possible to add a new option, to change the value of some, but it's not possible to delete an option defined in the bundled file.

When IJent WSL FS is enabled in an IDE by default, there are specific options in the bundled VMOptions file. This commit brings a way to disable these options without removing them from the bundled vmooptions file.

GitOrigin-RevId: 70589dab1292a06efc9bea207357206e5b783b84