[json] IJPL-157032 Disabled merging test
To fix it, we should make MatchResult store schemas in a set to avoid equal objects from being present in the resolve result. But this for some reason breaks ancient anyOf validation. The related code seems really hard to debug, so the merging test is currently disabled. The corresponding change is reverted in the previous commit.

GitOrigin-RevId: 08e7fd02cb6556044c3a80863bd29fa1d770155f