[terminal] IJPL-102728 Do not allow to undo command generation result
Since command is generated under `Command Generation` CommandProcessor's command, it is possible to undo it, but it is performed char by char.
The ideal solution would be to return to the Command Generation mode with the same typed prompt, but it requires a lot of changes.
So, better to just prohibit Undo for now.

GitOrigin-RevId: d26bbe2499a0369a7c755f627bcf1e6a5ade37e1