[Workspace Model] [IJPL-14579] Do not generate roots changed for maven projects after iml loading
This was the case for maven because of changes in `LibraryEntity` and `ContentRootEntity`.

`LibraryEntity` had change in the order of roots because we sort them before saving to the `.iml`. Now maven generates the roots in the sorted order. However, it's not clear if other subsystems are affected. This problem is discussed internally: https://youtrack.jetbrains.com/issue/IJPL-14579/There-is-a-progress-indicator-with-Scanning-files-to-index...-in-the-second-IDE-launch-warm-launch#focus=Comments-27-9988247.0-0

`ContentRootEntity` had changed because of children: new `SourceRootOrderEntity` was added after loading `iml` files. There is no need to generate a rootChanged event when the child of `ContentRootEntity` is updates, so a new rule was added about that.


(cherry picked from commit 6a2b4fc620ff3aeab4df9ffca9179f79bd61914f)

IJ-CR-137879

GitOrigin-RevId: a7a4c77956d511a12c00209827c1d8622d55baa4