IJPL-156653 Reset tabs scroll bar activity when selecting a tab
The problem happens in ScrollableSingleRowLayout.doScrollToSelectedTab,
namely in this call in the first line: tabs.isRecentlyActive().
What it's supposed to check is whether there is any recent scrolling activity,
and if there is, to not scroll to the currently selected tab to not interrupt
the scrolling activity.

However, the scrolling activity is detected by mouse movement and
scroll bar movement, and the latter happens when we just need to
switch to some tab, by clicking on it, by using a shortcut or by
selecting a tab from the hidden tab list. In all these cases
the scrolling activity is invoked by code, and should not
prevent actually scrolling to the active tab.

In all these cases the code that invokes the actual scrolling
is in EditorTabs.checkActive. That revalidateAndRepaint()
call there is what performs the actual scrolling.
Therefore, in order to fix this incorrect behavior,
it's enough to explicitly reset the scrolling activity flag there.
But since the activity tracking happens in the parent class,
we need an @Internal method for that.

Another way to solve this would be to not set the activity
flag at all unless there's actual user activity that involves
scrolling, but it seems much harder to implement,
as there's no obvious way to figure out what caused
the scroll bar model to change its value by the time
the event arrives. And the code that sets the activity
flag based on the mouse movement would also have
to be made much more complicated, as it would have
to not just track the mouse position, but somehow
figure out what the user is doing. For these reasons,
it seems more reasonable to just reset the flag
whenever we need to scroll ASAP.

GitOrigin-RevId: 1502afa4b6334b2f68a0f37fd13101d7d2894f60