[gradle][kotlin] IDEA-354292 prevent flaky error appearing for K2 test when test project caches are not created yet
Previously, a flaky error was reproducible locally by:
- deleting the test project caches - `rm -r system/test`
- running K2GradleTaskRunConfigurationProducerTest

Test was failing for each next Gradle version incoming as a parameter. If a test was already finished (successfully or not) for any Gradle version, it would not fail for the same version, even with other different test parameters. But if the test runs with any Gradle version for the first time, it could fail.

Most probably, it was failing because the test was starting before completing some K2 logic with a project in ProjectGradleSettingsListener in a coroutine. This change allows waiting for the coroutine completion before running a test.

GitOrigin-RevId: abb5f150f888e73708df7c1a17e7dcad79ba09e1