IJPL-149878 IJent WSL FS: module with constants, functions to enable in particular IDEs
To enable the new filesystem, some tricky VM options must be specified. Just changing the default value of the registry key is not enough.

The new code brings tooling and constants to ease enabling the IJent WSL filesystem in specific products.

`IjentWslNioFsToggler` stops specifying `-Didea.io.use.nio2` because `com.intellij.openapi.util.io.FileSystemUtil.computeMediator` is already aware of the IJent WSL filesystem and starts to use java.nio without the explicit system property.

`IjentWslNioFsToggler` keeps reacting to the registry key changes, but forces its value into the VM options, literally ignoring the state of the registry. It was done on purpose to make the VM options file the only source of truth and to not deal with registry-vmoptions synchronization issues.

GitOrigin-RevId: 6a3c333f841f87cece56ddc26bcbc284cf1a1ecf