IJPL-156860 Hide splash right when editor or any tool window is shown
Wait for these events during project loading, but not block it. If no editor or tool window is opened during project loading, then the splash will be hidden by the existing logic in the constructor of `IdeGlassPaneImpl`.
Also, reopen editors now in the loading scope. Otherwise, splash can be hidden before editor loading because of the logic in `IdeGlassPaneImpl` constructor in the case when there is no tool window.

GitOrigin-RevId: 2e942d0f24303392aa221521c21e3795d2052c7f