[settings-sync] IJPL-13201 Allow to choose cross-IDE sync mode when enabling sync
To facilitate re-using the sync scope selector part of the sync settings UI in the modal, it was extracted
to  the `SettingsSyncPanelFactory` helper (like the category UI) and exposed as a combined factory method
named `createCombinedSyncSettingsPanel`. The combined panel also took the responsibility for applying the
changed settings as not to duplicate logic. Use of sync scope state (that is, `SettingsSyncLocalSettings`)
was refactored to follow the same pattern as `SettingsSyncSettings` with peristent setting and an UI model,
both implementing a shared interface.

To actually apply the selected sync scope on enabling sync the `SettingsSyncBridge#applyInitialChanges`
method now creates/removes the cross-IDE sync server-side marker file before any sync data are sent
or applied. We also ensure we only touch this file if the cross-IDE sync state is supposed to change, as
treating those methods as upserts seem to cause spurious sync issues.  We also now force flushing settings
to disk at the start of `SettingsSyncBridge#initialize`, to ensure that synced data is fresh (otherwise the
first sync might end up sending stale data).