[gitlab] fix changes selection loop
The crux of the issue was that the data was incorrect.
UI events should be handled synchronously until the "single source of truth" is set and then that state can be propagated asynchronously.
That was not the case for tree/diff selection handlers. Selection updates were handled asynchronously which sometimes caused an updated loop while trying to synchronize two states.

#IJPL-77514 Fixed

GitOrigin-RevId: fa0e860bd893f10e09e11af49547b3844c31fd35