[json] IJPL-156923 Cache element validity against a particular schema node to avoid double computation
With the introduction of if-then-else resolve that selects a branch depending on a given instance file element, the number of calls to JsonSchemaResolver.isCorrect method increased significantly. It is expected, as any validation/reference contribution/other editor features require a resolved schema node.

Resolve results are cached on language modification tracker because with the instance file updates, a previously valid if-else branch might become invalid

GitOrigin-RevId: f197203a3c53a7c4e7c1d502faaff3d50a9451e3