Move Kotlin DataFrame AI action into shared dropdown.
Refactoring AI actions into a single dropdown caused some issues with Kotlin Notebooks where now both Python and Kotlin actions are shown and Kotlin actions are not in the dropdown.

This MR refactors a few things so everything is now in a single dropdown and we also check for the notebook language before showing the action so Kotlin DataFrames do not show Python AI actions and vice versa.

Merge-request: IJ-MR-137877
Merged-by: Christian Melchior <christian.melchior@jetbrains.com>

GitOrigin-RevId: 3eafb409f42cea6b85f26c3551315f075ed2df5e