IJPL-155732: consider plugin alias for module graph
Before, "plugin alias" (`<module name=""/>` tags in XMLs) were not considered correctly when calculating the initial set of module dependencies, and the dependencies for plugin/module might be later incorrectly considered as not satisfied.

Note that the v2 module (containing an alias) still has `CODE_ID` plugin id, so it also has to be handled.

GitOrigin-RevId: 06af2c24712823c987a08b51ede47bdc89eb0f56