IJPL-157271: indexing should not cache files content
Make things simple and consistent: most clients that pass "false" here (just like indexes) expect that content will NOT be cached (just like indexes). Almost all the places (besides indexes) where this method is invoked with "false" argument are working with binary files. The behavior of the code which invokes this method with the `true` argument does not change.

GitOrigin-RevId: 257573a46abf9a7d3fa163ccc8e6515c1d6045b2