[terminal] IJPL-156325 Refactor: use enum for storing prompt style setting and extract it to the separate service
Enum is more suitable here since we plan to add the 3rd option - single line prompt.
Also, it is worth extracting this setting to the separate PersistentStateComponent just to not mix the New Terminal settings with the classic ones.
Migration of the existing `useShellPrompt` setting is provided. So, if the user enabled the Shell prompt before this change, it will be still enabled after this change. In a release, we will be able to remove the old setting.


(cherry picked from commit 66cc0ae31f9932166ef4b43bf1b02c70f4b1954e)

IJ-CR-138338

GitOrigin-RevId: 3b7425c851594a4a5719b679d482a4874b3481b8