[terminal] IJPL-155973 Do not allow prompt text and action toolbar intersection
Give the action toolbar the separate space by decreasing the terminal width. The width of the output grid should be decreased as well to place the right part of the prompt at the same place in the output as in the prompt.


(cherry picked from commit 3a1a8f65730071977e8f03b343f854914cbd80ab)

IJ-CR-138539

GitOrigin-RevId: 598ad1b2ddb21626b55504ad92510ae9ad7cb084