[gitlab] Disable auto-closing JSON parsers to avoid premature closure (IJPL-157434)
#IJPL-157434 Fixed

This should prevent READING_LENGTH exceptions caused by
JsonParsers closing the reader before reading all tokens.

Also, don't check for readiness. `ready` depends on `java.io.InputStream.available`.
But, `available` does not block to check a correct count of available bytes, so
it might return 0 even if more bytes are available still.
Probably best to read to completion regardless of readiness to avoid these errors...

Interesting note: if we use `.factory` instead of the method on `mapper`
the configurations from `genericConfig` will not be applied to the parser created.

GitOrigin-RevId: 99ff0728d1c33f59e04cf33c5c34e452b8c41d97