IJPL-157535 Don't interpret layout changes in tabs as scroll bar activity
When the size of the tabs component changes, doLayout()
is called which calls updateScrollBarModel(), which changes
the scroll bar value, which in turn is detected as scroll bar activity.
This prevents the tabs from fully updating their layout.

The first idea to fix this was to store the tabs component size
when activity is detected. Then, later, when laying out the component,
compare the actual size to the stored one, and if it has resized,
ignore the activity flag and do full relayout. It didn't work
because the relayout operation itself is considered an activity,
and by the time it happens, the size is already new, so there's
nothing to compare with.

Instead, fix this by completely suspending the activity tracker
when laying out the component.

GitOrigin-RevId: 587d3709fdddff947377bdfcfb8ddff5353a3f3c