[RDCT] Prohibit saving settings under a foreign ClientId
Settings that are saved under a remote ClientId are not saved on disk (see `StateStorageBase::setState`) to support in-memory per-client settings. Before this commit calling `saveSettings` under a non-local clientId was allowed and silently led to settings not being saved at all.
This commit prohibits such behavior and throws an exception (leading to a "Failed to save settings" notification and warning or error in logs).


IJ-MR-138540

GitOrigin-RevId: e4bc634a163e443e0ef98964b8a8d4183b9c6921