RUBY-20139: check the existence of the current locale for macOS when computing LC_CTYPE.
There was an issue when the locale, computed based on the region and the language, was present in the result of `Locale.getAvailableLocales()` but was not actually installed on the user machine (e.g. en_CY, a valid locale in Unicode CLDR but not present on macOS). This meant that the computed value of LC_CTYPE would be incorrect, which can throw off some logic that depends on it even if it's completely unrelated to the user language and region. For example, in the issue linked above, rspec tests with Japanese in spec names would not run if LC_CTYPE was set to an invalid value.

IJ-CR-138549

GitOrigin-RevId: dce175ea79481cd10fbb8163f68fdc8d5cad0556