[RDCT] GTW-8997: Synchronize Language and Region from the frontend to the backend
This commit adds `LocalizationStateService` to the whitelist of synchronized settings, as well as adds a custom synchronizer for the `RegionSettings`. These settings are implemented using `Preferences`, which are not supported by the standard settings synchronization, so a custom one is required.
Note: `RegionSettings.UPDATE_TOPIC` was refactored because topics with a `Runnable` listener class do not support lazy subscribers.


IJ-MR-138540

GitOrigin-RevId: bacc5e85d5de76725e220c2dc7799ed7b8ffd789