K2: Use KotlinScopeEnlarger for scope provider
K2 `analysisScope` of `KaSession` does not contain scopes provided by
`KotlinScopeEnlarger` for `GlobalSearchScope`. As a result, the
`KotlinScopeEnlarger` registered by Android Studio to support the
completion is not triggered when app has class symbols generated from
resources whose scope is supposed to be added by the
`KotlinScopeEnlarger`. This commit enlarges the search scope inside
`IdeKotlinByModulesResolutionScopeProvider` with the registered
`KotlinScopeEnlarger`.

^KTIJ-30502