IJPL-157535 Use invokeLater for JBTabsImpl validation on scrolling
Scrollbar value changes happen during doLayout() call.
The revalidate() call at that time has no effect because it marks
component as invalid, but because it happens during validation,
when the validation is over, it's marked as valid again.
This leads to outdated layout sometimes.

Fix by using invokeLater(), so by the time revalidate() is called
we're no longer validating.

GitOrigin-RevId: 833665ec4d165724ac8a47e65414c94d2e161f95