[terminal] IJPL-155973 Fix incorrect prompt width on terminal opening
Prompt string width depends on the terminal width we specify when starting the terminal process. Terminal width depends on the available prompt component width, which depends on side toolbar width.
Toolbar width is updated asynchronously depending on the number of containing visible actions. There can be GenerateCommand action if AI Assistant is enabled, so we need to postpone getting the terminal width required for terminal process start until the toolbar is updated with actual size.

GitOrigin-RevId: 52c4d26733bcd3bb81aa304e19b9b94201669384