[RDCT] GTW-8997: Implement dialogs for locale and region mismatch between frontend and backend
This commit implements a custom synchronization of locale and region. This is done because a general settings controlled-based synchronization can't support a precise control over these settings to implement a mismatch detection.

Also, four new dialogs are added:
1. RemDev case, language mismatch
2. RemDev case, region mismatch
3. CodeWithMe case, language mismatch
4. CodeWithMe case, region mismatch

For RemDev, the dialog has only the "Yes" button, and restart is inevitable. For CodeWithMe there are "Yes/No/Cancel" options, and only the frontend is restarted. Also, for RemDev we change settings on the backend, but for CodeWithMe we change language and region for the frontend, if possible.


IJ-MR-138540

GitOrigin-RevId: 23a542bfadbc0326d9522177d0391062afa1713f