[Workspace Model] [IJPL-14579] Do not throw rootsChanged in case the order of library roots groups has changed
The previous project model collected roots grouped by the root type. Because the order of keys doesn't exist in maps, the `iml` save sorted the groups alphabetically. And after the `iml` was loaded, this sorting didn't affect the library itself.

In case of `LibraryEntity`, the roots are stored in the flat list, so such sorting will cause `LibraryEntity` change. With an additional logic, we'll avoid an extra "roots changed" event

GitOrigin-RevId: b82823dd5b1a2a266bdda73b65ef51af6c4f5f24