IJPL-157491 Use the new bulk expand API for TreeUtil.restoreExpandedPaths
The main reason for this change is that the new API
actually checks for every path whether it's a leaf.
This prevents a bug when we first mark a path as expanded
and then later we add it, so it shows as collapsed
(because it's newly added), but can't be expanded either
because the tree thinks it's expanded. An example
of this is the filtering tree, which tries to re-expand
every expanded path after refiltering, even those paths
that don't exist because of that refiltering.

A nice side effect is a considerable performance improvement
in case there are many expanded paths.

GitOrigin-RevId: a0c41c33ab374e75cb385c72a796abe917e3fdad