[rdct] don't add a temporary synthetic project created by the frontend process to the recent project list (RDCT-1409)
It would be even better to completely override RecentProjectsManager in the frontend process since it seems that all its functionality is irrelevant there. However, its API is rather complicated, and it's already overridden in different IDEs, so a simpler approach with explicit marking Project instances as temporary is implemented here.

GitOrigin-RevId: 7ff4962e11bff286b82520862bb26a62a64b7907