[RDCT] GTW-8997: Remove `RegionSettings.flush` because it is not needed
Initially, `flush` was added because sometimes Region didn't synchronize, and I supposed that it didn't write to the disk. But it turned out that it was a concurrent modification from host and client settings pages, and monolith IDE also doesn't call `flush` before restarting, and it works well. So `flush` is not needed anymore.


IJ-MR-138540

GitOrigin-RevId: 1b8031a91fff0643316841e86e46f209b8dde425