IJPL-148224 Implement license list overflow design
It turned out a bit harder than it seems. The tricky part is that
now we can't just set a single border on the content panel
of the activate panel. In case of overflow, the border should be
much smaller, and the main border goes inside the scroll pane.

Solve it by setting the regular border on the activation code
and license server panels, and let the subscription panel
do its own border handling.

Another tricky part is that the subscription panel is wrapped
into a login prompt panel, so we have to set the border
for the prompt itself too.

GitOrigin-RevId: 61ab93678df87512055ecd4e451a4e07ec1371ec