IJPL-797 `intellij.platform.projectModel.impl` review internal API
Refactoring: remove package-level @Internal annotation from com.intellij.workspaceModel.ide.legacyBridge package and add it to individual classes instead. This is needed to add a non-internal class to the package (IJ-CR-137628). Also, since the package is split between this module and intellij.platform.projectModel, it's better not to have package-level annotations in it to avoid confusion.

GitOrigin-RevId: db90c9704800f24910b40de21596d648e331ae58