GTW-9120 `IdePopupEventDispatcher` gets removed from `IdePopupManager` because the popup's component is not shown in RD
See the previous commit for more info about the original problem.

The popup is shown on the frontend. Therefore, the backend component is not shown by default.
But in RD we have a special method `UIUtil.markAsShowing` which we use to mark components as showing.

However, this doesn't apply to standard `.isShowing` method, `UIUtil.isShowing` should be used instead.

GitOrigin-RevId: 4e4001f6102f8dd028face40f585d0aa468da341