[workspace model] register implementations of VirtualFileUrl in the cache without mentioning their classes (IJPL-409)
We have a custom serializer for `VirtualFileUrl` which is used for its implementations. In the IDE process we use `VirtualFileUrlBridge` implementation for all instances, and in the JPS build process this class is not available, and `VirtualFileUrlImpl` is used instead. To avoid mentioning the specific class name in the cache, `VirtualFileUrlManagerImpl` now exposes the implementation class via `virtualFileUrlImplementationClass`, which is registered in Kryo.

This also speeds up serialization a bit since we don't need to iterate over all instances of VirtualFileUrl in the storage.

GitOrigin-RevId: a66ca59f6999d7e5dc59f822c7f38e5700a2a329