[kotlin] K2 IDE: add possibility to specify a list of plugins which is safe to enable in K2 mode
Currently, if a plugin depends on Kotlin plugin in K2 mode and the plugin is not marked as k2-compatible, then it's going to be disabled. We have multiple plugins depending on Kotlin. A lot of them do not use resolve. So we can analyze the API they use and explicitly enable such plugins in K2.

The list of such plugins will be set in KTIJ-30546

^KTIJ-30545 fixed

GitOrigin-RevId: 2208e42905374fa6b6e409425c0cacf9cf396e14