IFT-575 Fix unexpected restore in Find in Files lesson
When we request `virtualFile`, we need to get the editor. But there can be the moment when the currently selected editor is null during new editor tab opening. It caused `NoTextEditor` exception and stop of the lesson.
Better to use more specific check here. But anyway, it looks like current behavior with `NoTextEditor` exception is not suitable and should be reconsidered.


(cherry picked from commit 5744726f8c0bf5089380f8417408a75ef125c45a)

IJ-CR-139296

GitOrigin-RevId: d3a53fc503ccb6ecb9635bc6f9cd37baced08faa