GTW-7608 Refactor actions on the HelpAndResourcesPanel (Welcome screen -> Learn)
Use the real presentation instead of the template one.
`expandActionGroup` updates presentations of each action and adds all actual presentations to `presentationFactory`

We can retrieve `presentation` from `presentationFactory` to use its text instead of `AnAction.templateText`

This is the correct way to get action's text as `templateText` might not be actual in the case of provided IDE/DataContext.

GitOrigin-RevId: e028a674de441fca4d3320ff10fcd90a23bbe7ce