GTW-7608 Refactor actions on the HelpAndResourcesPanel (Welcome screen -> Learn)
Right now `WhatsNewAction` is not visible on the Welcome Screen because the template presentation of this action doesn't provide `text` (it's null).
The `text` for the action is set inside action update, so we should call `update` before showing the action.

Calling `update` action directly is prohibited, the action should be a part of a group.
So, we can move the action to `WelcomeScreen.LearnIdeHelp` instead of creating it manually.
In this case, the action will be returned from `expandActionGroup` with an updated presentation

GitOrigin-RevId: 2228d92b084060999357b351b16fb00cf6a2a33b