CWM-9511 Invoke `beforeShow` only for popups opened under the controller session
The method notifies `IdePopupManager` about a new opened popup.
The manager is not per-client, therefore, if `closeAllPopups` method closes all opened popups (doesn't matter if they were open on host or guest).

Also, the manager closes popups on clicks or input events (e.g. `StackingPopupDispatcherImpl.dispatchMouseEvent)`.
It leads to the issue when host's event closes the popup on a guest, as it is recognized as a click outside a popup.

Right now, the simplest way is to avoid notifying the manager about guest's popups.

GitOrigin-RevId: 96eebeb262d0d87b00ae3cf22bbb66d6e27a2f2d