[split] don't change the value of 'isConfirmExit' option in remote development sessions (RDCT-1505)
IdeLifecycleUiCustomization service is introduced to allow suppressing the default dialog asking for confirmation on exit without modifying the option. This way we ensure that users won't be able to change the value of this option manually after it set in code in remote dev mode. Also, running the IDE in remote dev mode won't affect settings for monolith (full) IDE.

GitOrigin-RevId: 051e8a20151e0bb584c5435b87640920121f100a