[RDCT] Add missing java types to `JsonHelper` class
When `OldLocalValueSupplierTag` is created for `EditorSettings.languageBreadcrumbsMap`, the `valueClass` in `valueToJson` method is `java.lang.Boolean`. Before this commit it was serialized to `JsonLiteral(isString = true, value)`, but it should be serialized to `JsonLiteral(isString = false, value)`. This did break settings synchronization in RDCT because the same values were `oldValue.equals(newValue) == false`.
This commit adds Java boxed primitive classes to the list.

GitOrigin-RevId: 485990a7de550f2aa091aa0329666b3c2f56c0ed