[workspace model] properly handle cases when iml file of a module is moved to another directory (IJPL-8707)
The common logic which updates VirtualFileUrl instances doesn't work here because the entity source for a module points to the parent directory of its iml file (see IJPL-158284). So we need to manually check this case, and also ensure that the path in ModuleStore is updated.


(cherry picked from commit cc518f86a9f22c4d9d57932edcb0479982907adb)

IJ-CR-139912

GitOrigin-RevId: 5b04293515a3233f91852420b7b3610407452cf2